purpose built `TypeLookup` type
